+++ setup.cfg
@@ -1,0 +1,26 @@
+[metadata]
+name = fetchtastic
+version = 0.1.0
+author = Jeremiah K
+author_email = jeremiahk@gmx.com
+description = Meshtastic Firmware and APK Downloader
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/jeremiah-k/fetchtastic
+license = MIT
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
+install_requires =
+    requests
+    pick
+    PyYAML
+    urllib3
+
+[options.entry_points]
+console_scripts =
+    fetchtastic = app.cli:main
